requirements.txt
boto3==1.12.31
botocore==1.15.36
dj-database-url==0.5.0
Django==1.11.24
django-bootstrap-tags==1.2.0
django-forms-bootstrap==3.1.0
django-shortcuts==1.6
django-storages==1.9.1
django-tinymce==2.6.0
docutils==0.15.2
env==0.1.0
gunicorn==20.0.4
heroku==0.1.4
jmespath==0.10.0
Pillow==7.1.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
sqlparse==0.3.1
stripe==2.44.0
